Adds the SetLineSpacing method to the Paragraph class. This method can be used to set or remove line spacing from a paragraph.
One method uses two parameters: the type of spacing and the amount of spacing (float). The other method uses one parameter: the type of spacing for which to set the value to auto; can be either before, after or both. The second method can also be used to remove spacing by using the type None.

Patch provided by Annika89. Great work!