Update Container.cs
The paragraph startindex and endindex are corrupted in Document.paragraphs[].
The result was a wrong insertion of paragraphs and tables. (at wrong locations)
The function GetParagraphs had the bug. the variable index in that function should be updated after creation of the paragraph not before

The function public virtual List<int> FindAll(string str, RegexOptions options) was only resulting one item in the list. instead of a list of find idexes 
If there was only find one no big problem but if it finds more the return index was wrong also .
I think  it is a  typo error, instead of  indexes[0]  it should be indexes[i].